// projeto-legado/src/screens/NewSurveys/Modals/SelectModal/Products/interface.ts
import { DataInterface } from '../../../manager/interface'
import { selectedForm } from '../../create/components/content/components/forms'

import { RelationProps } from './SelectComponent/interface'

type TabsNames =
  | 'Produto'
  | 'Fabricante'
  | 'Marca'
  | 'Categoria'
  | 'Sub-Categoria'
  | 'Linha de Produto'

export type AddedItemProps = {
  [name in TabsNames]: {
    items?: RelationProps[]
  }
}

export interface ModalProps {
  edit?: boolean
  item?: DataInterface
  createSelected?: AddedItemProps
  creteSetSelected?: React.Dispatch<React.SetStateAction<selectedForm>>
  reload?: () => void
}